requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0


# Bleeding edge Django
django==1.11.4 # pyup: >=1.10,<1.11

# Configuration
django-environ==0.4.4

# Models
django-model-utils==3.0.0

# Python-PostgreSQL Database Adapter
psycopg2==2.7.3

# Time zones support
pytz==2017.2

# Redis support
django-redis==4.8.0
redis>=2.10.5

# Sendgrid
sendgrid==3.6.5
sendgrid-django==3.0.0

# Password storage
argon2-cffi==16.3.0

djangorestframework==3.8.2
django-cors-headers==2.4.0
